feat(resize): resize an expanded stage
This commit adds the feature "resize". This allows to resize
expanded stages (plan items, discretionary items) and the case
plan model.

Therefor the CmmnLayouter uses the ManhattenLayout to layout a
connection (onPart or discretionary connection) on resize.

Furthermore the di of attached criteria is also updated when
resizing a stage.

Closes #27